requirements.txt
django
unidecode
openmeteo-py
python-dateutil
geopy